// src/interfaces/ISso.ts
export interface ISSOExchangeTokenResponse {
  access_token: string;
  token_type: string;
  expires_in: number;
  scope: string;
  jenis_kantor: string;
  keterangan_group_id: string[];
  path_foto: string;
  kode_jabatan: string;
  nama_lengkap: string;
  phone2: null;
  nama_grade: string;
  nama_unit_org: string;
  phone1: string;
  nik: string;
  kode_unit_org: string;
  nama_unit_kerja: string;
  nik_atasan_1: string;
  group_id: any[];
  kode_unit_kerja: string;
  has_kasir: null;
  flag_status: string;
  kode_grade: string;
  granted_app: any[];
  nama_jabatan: string;
  email: string;
  tanggal_lahir: Date;
  jti: string;
}